backend/requirements.txt
fastapi
uvicorn
gunicorn
tensorflow-cpu>=2.15.0
numpy<2.0.0
Pillow>=10.0.0
python-multipart
google-generativeai